forwarder/requirements.txt
azure-storage-blob==12.19.0
google-auth==2.23.4
requests==2.31.0
pyyaml==6.0.1
pydantic==2.5.2
azure-data-tables==12.5.0
azure-identity==1.15.0
